pyproject.toml
[tool.poetry]
name = "victoria"
version = "0.8.3"
description = "Automation toolbelt -- a single command with multiple pluggable subcommands for automating any number of 'toil' tasks that inhibit productivity."
authors = ["Sam Gibson"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
click = "^7.0.0"
marshmallow = "^3.2.1"
pyyaml = "^5.1.2"
appdirs = "^1.4.3"
azure-storage-blob = "^12.3.0"
cryptography = "^2.8.0"
azure-keyvault = "^4.0.0"
azure-identity = "^1.3.0"
dpath = "^2.0.1"
victoria-config = {path = "example_plugins/victoria_config"}
victoria-store = {path = "example_plugins/victoria_store"}
victoria-encrypt = {path = "example_plugins/victoria_encrypt"}
adal = "^1.2.4"
msrestazure = "^0.6.4"
azure-cli-core = "^2.18.0"

[tool.poetry.dev-dependencies]
setuptools = "^49.2.0"
pylint = "^2.5.3"
yapf = "^0.30.0"
pytest-cov = "^2.10.0"
pyfakefs = "^4.1.0"
munch = "^2.5.0"
sphinx = "^3.1.2"
sphinx-rtd-theme = "^0.5.0"
m2r = "^0.2.1"
sphinx-autoapi = "^1.4.0"
bandit = "^1.6.2"
coverage = "^5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
